+++ app/views/account/articles/_form.html.slim
@@ -1,6 +1,6 @@
 .row
   .medium-6.columns
-    = simple_form_for article, url: url do |f|
+    = simple_form_for [:account, article] do |f|
 
       legend
         | Create your article
