+++ app/views/account/articles/edit.html.slim
@@ -1,0 +1,1 @@
+= render "form", article: article
