+++ app/views/account/articles/index.html.slim
@@ -1,0 +1,7 @@
+.row
+  .columns
+    h2 User Articles
+    - articles.each do |article|
+      p = article.title
+      p = link_to "Edit", edit_account_article_path(article)
+      p = link_to "Delete", account_article_path(article), method: :delete, data: { confirm: "Are you sure?" }
