+++ app/views/account/articles/new.html.slim
@@ -1,0 +1,1 @@
+= render "form", article: article
