--- app/views/author_articles/edit.html.slim
@@ -1,1 +1,0 @@
-= render "form", article: article, url: author_article_path(article)
