--- app/views/author_articles/new.html.slim
@@ -1,1 +1,0 @@
-= render "form", article: article, url: author_articles_path
