--- app/views/author_articles/show.html.slim
@@ -1,3 +1,0 @@
-= render "articles/article"
-p = link_to "Edit", edit_author_article_path(article)
-p = link_to "Delete", author_article_path(article), method: :delete, data: { confirm: "Are you sure?" }
