+++ spec/features/account/articles_spec.rb
@@ -1,13 +1,13 @@
 require "rails_helper"
 
-feature "Articles" do
+feature "Account Articles" do
   include_context "current user signed in"
 
   scenario "Author shows own articles" do
     article_1 = create(:article, title: "First article", user: current_user)
     article_2 = create(:article, title: "Second article")
 
-    visit author_articles_path
+    visit account_articles_path
 
     expect(page).to have_text(article_1.title)
     expect(page).not_to have_text(article_2.title)
@@ -17,7 +17,7 @@ feature "Articles" do
   scenario "Author shows own article" do
     article = create(:article, user: current_user)
 
-    visit author_article_path(article)
+    visit account_article_path(article)
 
     expect(page).to have_text(article.title)
     expect(page).to have_text(article.body)
@@ -25,7 +25,7 @@ feature "Articles" do
   end
 
   scenario "Author creates article with valid params" do
-    visit new_author_article_path
+    visit new_account_article_path
 
     fill_form(:article, attributes_for(:article).slice(:title, :body))
     click_on submit(:article)
@@ -34,7 +34,7 @@ feature "Articles" do
   end
 
   scenario "Author creates article with invalid params" do
-    visit new_author_article_path
+    visit new_account_article_path
 
     click_on submit(:article)
 
@@ -45,7 +45,7 @@ feature "Articles" do
   scenario "Author updates article" do
     article = create(:article, title: "First article", user: current_user)
 
-    visit edit_author_article_path(article)
+    visit edit_account_article_path(article)
     fill_in "Title", with: "Updated title"
     click_on submit(:article, :update)
 
@@ -56,7 +56,7 @@ feature "Articles" do
   scenario "Author deletes own article" do
     article = create(:article, title: "First article", user: current_user)
 
-    visit author_article_path(article)
+    visit account_article_path(article)
     click_on("Delete")
 
     expect(page).to have_text("Article was successfully destroyed.")
